pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.8.0",
    "homeharvest>=0.8.9",
    "opyncorporates>=0.1",
    "usaddress>=0.5.16",
    "usps-api>=0.5",
]
